(MAINT) Move Platforms hash to the correct location
The Pkg::Platforms hash was defined directly under the Pkg:: namespace,
but stored under the pkg/config namespace on disk.
This corrects that by moving the file and updating any "require" calls
for pkg/config/platforms to pkg/platforms.
