+++ lib/packaging/platforms.rb
